gemini_analysis_function-source/requirements.txt
functions-framework==3.*
google-cloud-vision==3.4.4
vertexai==1.43.0
firebase-admin==6.4.0
